Refactor to commands
