+++ tfworker/commands/__init__.py
@@ -1,0 +1,3 @@
+from .base import BaseCommand  # noqa
+from .root import RootCommand  # noqa
+from .terraform import TerraformCommand  # noqa
