+++ tfworker/commands/base.py
@@ -1,0 +1,2 @@
+class BaseCommand:
+    pass
